requirements.txt
pymongo==3.12.0
django
-e .
dnspython
PyYAML
gunicorn
psycopg2
django-phone-field
